finish Step9. Redirection
