+++ log/error.txt
@@ -1,0 +1,1 @@
+error
